+++ api/src/main/java/io/gatehill/imposter/script/ResponseBehaviour.java
@@ -14,5 +14,7 @@ public interface ResponseBehaviour {
 
     String getResponseData();
 
+    String getExampleName();
+
     ResponseBehaviourType getBehaviourType();
 }